add toJson method to serialize the model
